move enums into their own file
