+++ src/components/settings/components/customAccentColor.tsx
@@ -13,8 +13,8 @@ import purpleEmoji from '../../../assets/accent-colors/purple.svg';
 import prideEmoji from '../../../assets/accent-colors/rainbow.svg';
 import redEmoji from '../../../assets/accent-colors/red.svg';
 import yellowEmoji from '../../../assets/accent-colors/yellow.svg';
-import {BetterTweetDeckAccentColors} from '../../../features/themeTweaks';
 import {HandlerOf} from '../../../helpers/typeHelpers';
+import {BetterTweetDeckAccentColors} from '../../../types/btdSettingsEnums';
 import {Trans} from '../../trans';
 import {generateInputId} from '../settingsHelpers';
 import {BaseSettingsProps} from '../settingsTypes';
