+++ src/components/settings/components/themeSelector.tsx
@@ -4,8 +4,8 @@ import React, {ReactNode} from 'react';
 import defaultDarkTheme from '../../../assets/dark-themes/default-dark.png';
 import lightsOut from '../../../assets/dark-themes/lights-out.png';
 import oldDark from '../../../assets/dark-themes/old-dark.png';
-import {BetterTweetDeckThemes} from '../../../features/themeTweaks';
 import {HandlerOf} from '../../../helpers/typeHelpers';
+import {BetterTweetDeckThemes} from '../../../types/btdSettingsEnums';
 import {Trans} from '../../trans';
 import {generateInputId} from '../settingsHelpers';
 import {featureBadgeClassname, NewFeatureBadge} from './newFeatureBadge';
