+++ src/components/settings/menu/settingsGeneral.tsx
@@ -6,8 +6,8 @@ import heartIcon from '../../../assets/heart-icon.svg';
 import arrowsIcon from '../../../assets/mutual-icon.svg';
 import translatorIcon from '../../../assets/translator-icon.svg';
 import verifiedIcon from '../../../assets/verified-icon.svg';
-import {BTDMutualBadges} from '../../../features/badgesOnTopOfAvatars';
 import {isFirefox, isSafari} from '../../../helpers/browserHelpers';
+import {BTDMutualBadges} from '../../../types/btdSettingsEnums';
 import {BTDSettings} from '../../../types/btdSettingsTypes';
 import {Trans} from '../../trans';
 import {CheckboxSelectSettingsRow} from '../components/checkboxSelectSettingsRow';
