+++ src/features/themeTweaks.ts
@@ -1,24 +1,7 @@
 import './lightsOut.css';
 
-export enum BetterTweetDeckAccentColors {
-  DEFAULT = 'rgb(29, 161, 242)',
-  YELLOW = 'rgb(255, 173, 31)',
-  RED = 'rgb(224, 36, 94)',
-  PINK = 'rgb(224, 36, 142)',
-  PURPLE = 'rgb(121, 75, 196)',
-  ORANGE = 'rgb(244, 93, 34)',
-  GREEN = 'rgb(23, 191, 99)',
-  CUSTOM = 'CUSTOM',
-}
-
-export enum BetterTweetDeckThemes {
-  LIGHT = 'light',
-  DARK = 'default',
-  LEGACY_DARK = 'legacy',
-  ULTRA_DARK = 'ultra_dark',
-}
-
 import {makeBTDModule} from '../types/btdCommonTypes';
+import {BetterTweetDeckAccentColors, BetterTweetDeckThemes} from '../types/btdSettingsEnums';
 
 export const tweakTweetDeckTheme = makeBTDModule(({settings}) => {
   if (settings.customAccentColor !== BetterTweetDeckAccentColors.DEFAULT) {
